+++ solver_singles.py
@@ -30,7 +30,7 @@ def v_iter_single(setup,t,EV,female,ushift,force_f32=False):
     R = setup.pars['R_t'][t]
     
     
-    dtype_here = np.float32 if force_f32 else dtype
+    dtype_here = np.float64 if force_f32 else dtype
 
     
     
@@ -63,7 +63,7 @@ def v_iter_single(setup,t,EV,female,ushift,force_f32=False):
 #    money = wf + wm + asset_income
 #    aaa=money-c_opt-x_opt-s_opt
     
-    
+   
     EVexp = setup.vsgrid_s.apply_preserve_shape(EV)
     V_ret = setup.u_single_pub(c_opt,x_opt,ls) + ushift + beta*np.take_along_axis(EVexp,i_opt,0)
